refactor: remove resolution-epic dead code and document the new resolution surfaces
Post-epic cleanup for the FeatureGroup resolution work, in two parts: remove dead
code left after the epic, then document the surviving resolution surface.

Dead code removed:
- FeatureChainParser._match_pattern_based_feature (only self-referenced).
- format_feature_group_classes (plural) and its dedicated test file; the singular
  format_feature_group_class stays and keeps direct unit tests in a dedicated file.
- the log_n_subclasses parameter and its unreachable debug branch in get_all_subclasses.
- Options.add, a zero-caller legacy shim.

Behavior-preserving: get_column_base_feature now consumes the exported COLUMN_SEPARATOR.
The _strict_validation_rejection_reason / name_path_presence_rejection_reason diagnostic
seam is documented as supported (docstring-only, no behavior change).

Docs: typed FeatureResolutionError framing plus a catching example on the troubleshooting
page; diagnose / resolution_report() and the FeatureResolutionError / ResolutionDiagnosis /
ResolutionRecord types in mloda-api.md; get_result / get_artifacts docstrings.

Breaking changes (public API): Options.add is removed (use add_to_group); get_all_subclasses
loses its log_n_subclasses parameter. Both were zero-caller and intentional.
