+++ mloda/core/abstract_plugins/components/utils.py
@@ -49,13 +49,11 @@ def as_str(value: Any) -> str:
     return value
 
 
-def get_all_subclasses(cls: Any, log_n_subclasses: int = 0) -> set[type[Any]]:
+def get_all_subclasses(cls: Any) -> set[type[Any]]:
     all_subclasses = set()
 
     for subclass in cls.__subclasses__():
         all_subclasses.add(subclass)
         all_subclasses.update(get_all_subclasses(subclass))
 
-    if log_n_subclasses > 0:
-        logger.debug(f"Abstractclass: {type(cls)}. Subclasses: {list(all_subclasses)[log_n_subclasses]}.")
     return all_subclasses
